v2/requirements.txt
accelerate==0.33.0
datasets==2.21.0
torch>=2.2.0
transformers==4.44.2
evaluate==0.4.2
scikit-learn==1.5.1
numpy==1.26.4
tqdm==4.66.4
huggingface_hub==0.23.5
xformers>=0.0.27
pyarrow==17.0.0
pillow==10.4.0
requests==2.32.3
diskcache==5.6.3
